move protected/too many datasets error tracking functions to new dialog
