+++ context/app/static/js/components/workspaces/CreateWorkspaceDialog/CreateWorkspaceWithDatasets.jsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useRef } from 'react';
+import React, { useCallback } from 'react';
 import Button from '@mui/material/Button';
 
 import WorkspaceField from 'js/components/workspaces/WorkspaceField';
@@ -25,8 +25,6 @@ const errorHelper = {
 export default function CreateWorkspaceWithDatasetsDialog({ ...rest }) {
   const { selectedRows, deselectRows } = useSelectableTableStore();
   const { toastSuccess } = useSnackbarActions();
-  const reportedProtectedRows = useRef(false);
-  const reportedTooManyRows = useRef(false);
   const protectedRows = useDatasetsAccessLevel(selectedRows.size > 0 ? [...selectedRows] : []).datasets;
   const containsProtectedDataset = protectedRows.length > 0;
 
@@ -34,28 +32,12 @@ export default function CreateWorkspaceWithDatasetsDialog({ ...rest }) {
 
   if (selectedRows.size > 10) {
     errorMessages.push(errorHelper.datasets(selectedRows));
-    if (!reportedTooManyRows.current) {
-      reportedTooManyRows.current = true;
-      trackEvent({
-        category: 'Workspace Creation',
-        action: 'Too many datasets selected',
-        value: selectedRows.size,
-      });
-    }
   }
 
   if (containsProtectedDataset) {
     const protectedRowsError =
       protectedRows.length === 1 ? errorHelper.protectedDataset : errorHelper.protectedDatasets;
     errorMessages.push(protectedRowsError(protectedRows));
-    if (!reportedProtectedRows.current) {
-      reportedTooManyRows.current = true;
-      trackEvent({
-        category: 'Workspace Creation',
-        action: 'Protected datasets selected',
-        value: protectedRows.map((r) => r._id),
-      });
-    }
   }
   const protectedHubmapIds = protectedRows?.map((row) => row._source.hubmap_id).join(', ');
